+++ source/managers/ECSManager.inl
@@ -4,6 +4,6 @@ template <class S, class... Args>
 void brk::ecs::Manager::AddSystem(Args&&... args)
 {
 	BRK_ASSERT(s_SysIndex<S> >= m_Systems.size(), "System has already been registered!");
-	s_SysIndex<S> = m_Systems.size();
+	s_SysIndex<S> = static_cast<uint32>(m_Systems.size());
 	m_Systems.emplace_back(SystemInstance::Create<S>(std::forward<Args>(args)...));
 }
